+++ nip/utils/plotting/__init__.py
@@ -1,0 +1,3 @@
+"""Utilities for plotting the results of NIP experiments."""
+
+from .decision_agreement import DecisionAgreementAnalysis, analyse_decision_agreement